# code/docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: 'REST API'
  version: 1.0.0
paths:
...
  '/v1/notifications/{id}':
    put:
      tags:
        - Notifications
      summary: 'Mark notification as read'
      description: 'Marks a specific notification as read for the authenticated user'
      operationId: 4865d1628fb7f7e5f28e62d89989d694
      parameters:
        -
          name: id
          in: path
          description: 'Notification ID'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Notification marked as read successfully'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: success }
                  message: { type: string, example: Ok }
                type: object
...
  /api/v1/profile/settings:
    get:
      tags:
        - Profile
      summary: 'Get settings'
      description: 'Get all user settings grouped by category'
      operationId: fa868c3419e45d10528e357cb7dc708d
      responses:
        '200':
          description: 'Returns user settings grouped by category'
          content:
            application/json:
              schema:
                properties:
                  data: { description: 'Settings data grouped by category', type: object }
                type: object
    put:
      tags:
        - Profile
      summary: 'Change settings'
      description: 'Updates settings based on provided parameters'
      operationId: 42286e8aa955d872524f5600156f96c7
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                settings:
                  type: array
                  items: { properties: { category: { type: string, enum: [GENERAL, ACCOUNT, NOTIFICATION], example: NOTIFICATION }, name: { type: string, enum: [phone_verified_at, email_verified_at, accepted_cookies], example: accepted_cookies }, value: { type: mixed, example: true } }, type: object }
              type: object
      responses:
        '200':
          description: 'Setting updated successfully'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string, example: 'Setting updated successfully' }
                type: object
        '400':
          description: 'Validation error'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: false }
                  message: { type: string, example: 'Validation failed' }
                  errors: { type: array, items: { type: string } }
                type: object
  /api/v1/profile/setting/verifications:
    post:
      tags:
        - Profile
      summary: 'Create verification request'
      description: 'Creates a new verification request and sends a 6-digit code'
      operationId: db5e8ecb9f53e0bb70ddafa0efb2aa8e
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                type:
                  description: 'Verification type'
                  type: string
                  enum: [email, phone]
              type: object
      responses:
        '200':
          description: 'Code sent successfully'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: success }
                type: object
  '/api/v1/profile/setting/verifications/{type}':
    put:
      tags:
        - Profile
      summary: 'Complete verification'
      description: 'Verifies and completes a verification request'
      operationId: 691de1108cbe54162ba0516e822fe581
      parameters:
        -
          name: type
          in: path
          description: 'Type of verification (email or phone)'
          required: true
          schema:
            type: string
            enum:
              - email
              - phone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                code:
                  description: 'Verification code'
                  type: string
                  maxLength: 6
                  minLength: 6
              type: object
      responses:
        '200':
          description: 'Code verified successfully'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: success }
                type: object
        '400':
          description: 'Invalid verification code'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: error }
                  message: { type: string, example: 'Invalid verification code' }
                type: object
  /api/v1/profile/user/password:
    put:
      tags:
        - Profile
      summary: 'Change user password'
      description: 'Changes the password for the authenticated user'
      operationId: e612eff67a2cf0133671fcb0503eb233
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - currentPassword
                - newPassword
              properties:
                currentPassword:
                  type: string
                  example: current123
                newPassword:
                  type: string
                  example: new123
              type: object
      responses:
        '200':
          description: 'Password changed successfully'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string, example: Ok }
                type: object
        '400':
          description: 'Validation error'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: false }
                  message: { type: string, example: 'Password change failed' }
                type: object
    post:
      tags:
        - Profile
      summary: 'Create user password'
      description: 'Create the password for the authenticated user'
      operationId: 68e4bb483437d99c91fbc985909b1fb8
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - password
                - confirmationPassword
              properties:
                password:
                  type: string
                  example: current123
                confirmationPassword:
                  type: string
                  example: new123
              type: object
      responses:
        '200':
          description: 'Password created successfully'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string, example: Ok }
                type: object
        '400':
          description: 'Validation error'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: false }
                  message: { type: string, example: 'Password created failed' }
                type: object
...